+++ sdk/healthbot/Azure.ResourceManager.HealthBot/CHANGELOG.md
@@ -1,10 +1,25 @@
 # Release History
 
-## 1.3.0 (2026-06-04)
+## 1.4.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
 
 ### Other Changes
 
-- This release contains the same features and fixes as 1.3.0-beta.1, now promoted to stable.
+## 1.3.0 (2026-06-13)
+
+### Features Added
+
+- Upgraded api-version to 2025-11-01.
+
+### Other Changes
+
+- Upgraded dependent Azure.Core to 1.59.0.
+- Upgraded dependent Azure.ResourceManager to 1.14.0.
 
 ## 1.2.0 (2025-10-22)